fix distortion issue with 3857
