+++ hydroengine_service/config.py
@@ -1,4 +1,14 @@
-#!/usr/bin/env python
+import pathlib
+import json
+
 """Required credentials configuration."""
 EE_ACCOUNT = '578920177147-ul189ho0h6f559k074lrodsd7i7b84rc@developer.gserviceaccount.com'
 EE_PRIVATE_KEY_FILE = 'privatekey.json'
+
+APP_DIR = pathlib.Path(__file__).parent
+DATASET_DIR = APP_DIR / 'datasets'
+DATASET_PATH = DATASET_DIR / 'dataset_visualization_parameters.json'
+
+# load visualization parameters
+with DATASET_PATH.open() as json_file:
+    DATASETS_VIS = json.load(json_file)
